+++ RELEASE_NOTES.md
@@ -103,6 +103,9 @@ IDEs, text editors, and syntax highlighters support
 * FIXED: Emacs syntax coloring support for using tab-based indentation as per
 coding guidelines but to default to user-defined tab width.
 
+* FIXED: Emacs syntax coloring support for the `create_logtalk_flag/2` built-in
+predicate.
+
 
 3.09.0 - November 28, 2016
 ==========================